Extract P3 field macro, add Mersene31Field
