+++ number/Cargo.toml
@@ -14,6 +14,7 @@ ark-bn254 = { version = "0.4.0", default-features = false, features = [
 ark-ff = "0.4.2"
 ark-serialize = "0.4.2"
 p3-baby-bear = { git = "https://github.com/powdr-labs/Plonky3.git", branch = "main" }
+p3-mersenne-31 = { git = "https://github.com/powdr-labs/Plonky3.git", branch = "main" }
 p3-field = { git = "https://github.com/powdr-labs/Plonky3.git", branch = "main" }
 num-bigint = { version = "0.4.3", features = ["serde"] }
 num-traits = "0.2.15"
